fix: do not overflow when computing partition size

+++ common/src/main/scala/org/neo4j/spark/config/TopN.scala
@@ -18,4 +18,4 @@ package org.neo4j.spark.config
 
 import org.apache.spark.sql.connector.expressions.SortOrder
 
-case class TopN(limit: Int, orders: Array[SortOrder] = Array.empty)
+case class TopN(limit: Long, orders: Array[SortOrder] = Array.empty)
